Updated engine sound, added more buses to the bus layout
The engine sound was made in FL Studio by recreating the V8 engine
firing pattern and running it really fast. It doesn't sound good
partly because a single sample is used instead of different samples
at different RPMs.

Not all the buses are going to be used in the end for the demo,
they're just an example